added new file but still have issue with the csrf tokening

+++ controllers/admin.js
@@ -1,5 +1,7 @@
 const mongoose = require("mongoose");
 
+const fileHelper = require("../util/file");
+
 const { validationResult } = require("express-validator/check");
 
 const Product = require("../models/product");
@@ -11,7 +13,7 @@ exports.getAddProduct = (req, res, next) => {
 		editing: false,
 		hasError: false,
 		errorMessage: null,
-		validationErrors: []
+		validationErrors: [],
 	});
 };
 
@@ -20,40 +22,39 @@ exports.postAddProduct = (req, res, next) => {
 	const image = req.file;
 	const price = req.body.price;
 	const description = req.body.description;
-	if (!image) {
-		return res.status(422).render("admin/edit-product", {
-			pageTitle: "Add Product",
-			path: "/admin/add-product",
-			editing: false,
-			hasError: true,
-			product: {
-				title: title,
-				price: price,
-				description: description
-			},
-			errorMessage: "Attached file is not an image.",
-			validationErrors: []
-		});
-	}
-	const errors = validationResult(req);
+	// if (!image) {
+	// 	return res.status(422).render("admin/edit-product", {
+	// 		pageTitle: "Add Product",
+	// 		path: "/admin/add-product",
+	// 		editing: false,
+	// 		hasError: true,
+	// 		product: {
+	// 			title: title,
+	// 			price: price,
+	// 			description: description,
+	// 		},
+	// 		errorMessage: "Attached file is not an image.",
+	// 		validationErrors: [],
+	// 	});
+	// }
+	// const errors = validationResult(req);
 
-	if (!errors.isEmpty()) {
-		console.log(errors.array());
-		return res.status(422).render("admin/edit-product", {
-			pageTitle: "Add Product",
-			path: "/admin/add-product",
-			editing: false,
-			hasError: true,
-			product: {
-				title: title,
-				imageUrl: imageUrl,
-				price: price,
-				description: description
-			},
-			errorMessage: errors.array()[0].msg,
-			validationErrors: errors.array()
-		});
-	}
+	// if (!errors.isEmpty()) {
+	// 	console.log(errors.array());
+	// 	return res.status(422).render("admin/edit-product", {
+	// 		pageTitle: "Add Product",
+	// 		path: "/admin/add-product",
+	// 		editing: false,
+	// 		hasError: true,
+	// 		product: {
+	// 			title: title,
+	// 			price: price,
+	// 			description: description,
+	// 		},
+	// 		errorMessage: errors.array()[0].msg,
+	// 		validationErrors: errors.array(),
+	// 	});
+	// }
 
 	const imageUrl = image.path;
 
@@ -62,17 +63,17 @@ exports.postAddProduct = (req, res, next) => {
 		title: title,
 		price: price,
 		description: description,
-		imageUrl: imageUrl,
-		userId: req.user
+		imageUrl: "", //imageUrl,
+		userId: req.user,
 	});
 	product
 		.save()
-		.then(result => {
+		.then((result) => {
 			// console.log(result);
 			console.log("Created Product");
 			res.redirect("/admin/products");
 		})
-		.catch(err => {
+		.catch((err) => {
 			// return res.status(500).render('admin/edit-product', {
 			//   pageTitle: 'Add Product',
 			//   path: '/admin/add-product',
@@ -89,7 +90,6 @@ exports.postAddProduct = (req, res, next) => {
 			// });
 			// res.redirect('/500');
 			const error = new Error(err);
-			console.log(err);
 			error.httpStatusCode = 500;
 			return next(error);
 		});
@@ -102,7 +102,7 @@ exports.getEditProduct = (req, res, next) => {
 	}
 	const prodId = req.params.productId;
 	Product.findById(prodId)
-		.then(product => {
+		.then((product) => {
 			if (!product) {
 				return res.redirect("/");
 			}
@@ -113,10 +113,10 @@ exports.getEditProduct = (req, res, next) => {
 				product: product,
 				hasError: false,
 				errorMessage: null,
-				validationErrors: []
+				validationErrors: [],
 			});
 		})
-		.catch(err => {
+		.catch((err) => {
 			const error = new Error(err);
 			error.httpStatusCode = 500;
 			return next(error);
@@ -127,7 +127,7 @@ exports.postEditProduct = (req, res, next) => {
 	const prodId = req.body.productId;
 	const updatedTitle = req.body.title;
 	const updatedPrice = req.body.price;
-	const updatedImageUrl = req.body.imageUrl;
+	const image = req.file;
 	const updatedDesc = req.body.description;
 
 	const errors = validationResult(req);
@@ -140,31 +140,33 @@ exports.postEditProduct = (req, res, next) => {
 			hasError: true,
 			product: {
 				title: updatedTitle,
-				imageUrl: updatedImageUrl,
 				price: updatedPrice,
 				description: updatedDesc,
-				_id: prodId
+				_id: prodId,
 			},
 			errorMessage: errors.array()[0].msg,
-			validationErrors: errors.array()
+			validationErrors: errors.array(),
 		});
 	}
 
 	Product.findById(prodId)
-		.then(product => {
+		.then((product) => {
 			if (product.userId.toString() !== req.user._id.toString()) {
 				return res.redirect("/");
 			}
 			product.title = updatedTitle;
 			product.price = updatedPrice;
 			product.description = updatedDesc;
-			product.imageUrl = updatedImageUrl;
-			return product.save().then(result => {
+			if (image) {
+				fileHelper.deleteFile(product.imageUrl);
+				product.imageUrl = image.path;
+			}
+			return product.save().then((result) => {
 				console.log("UPDATED PRODUCT!");
 				res.redirect("/admin/products");
 			});
 		})
-		.catch(err => {
+		.catch((err) => {
 			const error = new Error(err);
 			error.httpStatusCode = 500;
 			return next(error);
@@ -175,15 +177,15 @@ exports.getProducts = (req, res, next) => {
 	Product.find({ userId: req.user._id })
 		// .select('title price -_id')
 		// .populate('userId', 'name')
-		.then(products => {
+		.then((products) => {
 			console.log(products);
 			res.render("admin/products", {
 				prods: products,
 				pageTitle: "Admin Products",
-				path: "/admin/products"
+				path: "/admin/products",
 			});
 		})
-		.catch(err => {
+		.catch((err) => {
 			const error = new Error(err);
 			error.httpStatusCode = 500;
 			return next(error);
@@ -192,12 +194,19 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
 	const prodId = req.body.productId;
-	Product.deleteOne({ _id: prodId, userId: req.user._id })
+	Product.findById(prodId)
+		.then((product) => {
+			if (!product) {
+				return next(new Error("Product not found."));
+			}
+			fileHelper.deleteFile(product.imageUrl);
+			return Product.deleteOne({ _id: prodId, userId: req.user._id });
+		})
 		.then(() => {
 			console.log("DESTROYED PRODUCT");
 			res.redirect("/admin/products");
 		})
-		.catch(err => {
+		.catch((err) => {
 			const error = new Error(err);
 			error.httpStatusCode = 500;
 			return next(error);
